https://extract.atlassian.net/browse/ISSUE-13016 Edits to order picker column labels.
https://extract.atlassian.net/browse/ISSUE-12989
Added pre-populated LabDEPatient and LabDEOrder data to Demo_LabDE database.

https://extract.atlassian.net/browse/ISSUE-12814
Updated demo database to the latest schema.

https://extract.atlassian.net/browse/ISSUE-13022
LabDE version in user license once again shows the version number.

Fixed some compile and FXCop warnings.
Fixed some duplicate ELI codes.

Reviewed by:
Nat (0/0)

[git-vault-id] Extract$/Engineering@33255/95419

+++ ReusableComponents/COMComponents/UCLIDRasterAndOCRMgmt/Core/Code/SpatialString_StringOperations.cpp
@@ -881,7 +881,7 @@ STDMETHODIMP CSpatialString::RemoveText(ISpatialString* pTextToRemove, long nPag
 
 		// If all the required characters-to-remove were found, remove the characters in
 		// setLettersToRemove.
-		if (nFirstPosition != -1 && setLettersToRemove.size() >= nNumLettersRequiredToRemove)
+		if (nFirstPosition != -1 && setLettersToRemove.size() >= (size_t)nNumLettersRequiredToRemove)
 		{
 			long nNewLetterCount = nLetterCount - setLettersToRemove.size();
 			vector<CPPLetter> vecNewLetters;
